Fix compare model and training flow bugs.
- Fix training flow bug: Set proper end time. Must cut training set off
  to leave enough room for eval set.
- Fix circular imports
- Combine logic of fetching data for input to the model into its own
  function.
